requirements.txt
fastapi==0.72.0
passlib==1.7.4
bcrypt
python-multipart==0.0.5
SQLAlchemy==1.4.31
uvicorn==0.17.0
python-jose=3.3.0